Working concept with Auth0 and refactor
Upgraded packages.
Refactored js WIP to typescript.
Integrated with Auth0
Added ability to create projects only for logged in users; Users can only see the projects they administrate unless they are an admin of the site; columns and tickets can not be deleted.

Ideas still to implement after this proof of concept:
- Integration with Github (need to deploy this to a public url first).
- Project access control to share projects between users.
- Link tickets to github issues
- Assign tickets to users.